+++ lib/bindings/cstruct_stubs.c
@@ -15,6 +15,7 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
+#include "solo5.h"
 
 #include <string.h>