+++ lib/main.ml
@@ -21,70 +21,54 @@
  * 02111-1307, USA.
  *)
 
-open Lwt
+external solo5_yield : Time.t -> int64 = "mirage_solo5_yield_2"
 
-external riscv_yield : [`Time] Time.Monotonic.t -> bool = "mirage_riscv_yield"
+(* A Map from Int64 (solo5_handle_t) to an Lwt_condition. *)
+module HandleMap = Map.Make(Int64)
+let work = ref HandleMap.empty
 
-let exit_hooks = Lwt_dllist.create ()
-let enter_hooks = Lwt_dllist.create ()
-let exit_iter_hooks = Lwt_dllist.create ()
-let enter_iter_hooks = Lwt_dllist.create ()
-
-let rec call_hooks hooks  =
-  match Lwt_dllist.take_opt_l hooks with
-    | None ->
-        return ()
-    | Some f ->
-        (* Run the hooks in parallel *)
-        let _ =
-          Lwt.catch f
-          (fun exn ->
-            Printf.printf "ERROR: call_hooks(): Unhandled exception: %s\n%!" (Printexc.to_string exn);
-            return ()) in
-        call_hooks hooks
-
-(* riscv currently has an all-or-nothing interface to block and wait for I/O
- * events, so we use a single condition variable to block threads which are
- * waiting for work and wake them all up if I/O is possible.  This will need to
- * be extended once riscv_yield() gains support for selecting events of
- * interest. *)
-let work = Lwt_condition.create ()
-let wait_for_work () = Lwt_condition.wait work
-
-let err exn =
-  Logs.err (fun m -> m "main: %s\n%s" (Printexc.to_string exn) (Printexc.get_backtrace ())) ;
-  exit 1
+(* Wait for work on handle [h]. The Lwt_condition and HandleMap binding are
+ * created lazily the first time [h] is waited on. *)
+let wait_for_work_on_handle h =
+  match HandleMap.find h !work with
+    | exception Not_found ->
+      let cond = Lwt_condition.create () in
+      work := HandleMap.add h cond !work;
+      Lwt_condition.wait cond
+    | cond ->
+      Lwt_condition.wait cond
 
 (* Execute one iteration and register a callback function *)
 let run t =
-  let t = call_hooks enter_hooks <&> t in
   let rec aux () =
     Lwt.wakeup_paused ();
-    Time.restart_threads Time.Monotonic.time;
-    match (try Lwt.poll t with exn -> err exn) with
+    Time.restart_threads Time.time;
+    match Lwt.poll t with
     | Some () ->
         ()
     | None ->
+        (* Call enter hooks. *)
+        Mirage_runtime.run_enter_iter_hooks () ;
         let timeout =
           match Time.select_next () with
-          |None -> Time.Monotonic.(time () + of_nanoseconds 86_400_000_000_000L) (* one day = 24 * 60 * 60 s *)
+          |None -> Int64.add (Time.time ()) (Duration.of_day 1)
           |Some tm -> tm
         in
-        if riscv_yield timeout then begin
-          (* Call enter hooks. *)
-          Lwt_dllist.iter_l (fun f -> f ()) enter_iter_hooks;
+        let ready_set = solo5_yield timeout in
+        if not (Int64.equal 0L ready_set) then begin
           (* Some I/O is possible, wake up threads and continue. *)
-          Lwt_condition.broadcast work ();
-          (* Call leave hooks. *)
-          Lwt_dllist.iter_l (fun f -> f ()) exit_iter_hooks;
-          aux ()
-        end else begin
-          aux ()
-        end in
+          let is_in_set set x =
+            not Int64.(equal 0L (logand set (shift_left 1L (to_int x)))) in
+          HandleMap.iter (fun k v ->
+            if is_in_set ready_set k then Lwt_condition.broadcast v ()) !work
+        end;
+        (* Call leave hooks. *)
+        Mirage_runtime.run_leave_iter_hooks () ;
+        aux ()
+  in
   aux ()
 
-let () = at_exit (fun () -> run (call_hooks exit_hooks))
-let at_exit f = ignore (Lwt_dllist.add_l f exit_hooks)
-let at_enter f = ignore (Lwt_dllist.add_l f enter_hooks)
-let at_exit_iter f = ignore (Lwt_dllist.add_l f exit_iter_hooks)
-let at_enter_iter f = ignore (Lwt_dllist.add_l f enter_iter_hooks)
+let () =
+  at_exit (fun () ->
+    Lwt.abandon_wakeups () ;
+    run (Mirage_runtime.run_exit_hooks ()))